Powerpc64: Add POWER9 on the models
I currently booted FreeBSD on a POWER9 machine (on top of KVM), and I
got the following message:

cpu0: Unknown PowerPC CPU revision 0x1201, 2550.00 MHz

This is because of POWER9 is not listed as a valid CPU model. This patch
simply adds it with the same feature set as POWER8 for now, so, new
features could be added as they are implemented.