Powerpc64: Enable Hardware Transactional Memory
This patch enable Hardware Transactional Memory (HTM) facility.

The facility is only enabled on a process that executes a HTM instruction.  It
will raise and facility trap, which will enable flag the process with PCB_HTM,
from there after, the context switch will save and respore the 3 SPRs that are
associated with HTM.