powerpc64: Add DSCR support
Powerpc64 has support for a register called Data Stream Control Register
(DSCR), which basically control how the hardware controls the caching and
prefetch for stream operations.

Since mfdscr and mtdscr are privileged instructions, we need to emulate it, and
keep the DSCR configuration per thread.

The purpose of this feature is to change DSCR depending on the operation, set
to DSCR Default Prefetch Depth to deepest on string operations, as memcpy.