Powerpc64: Enable Hardware Transactional Memory
This patch enable Hardware Transactional Memory (HTM) facility.

The facility is only enabled on a process that executes a HTM instruction. It
will raise and facility trap, which will enable flag the process with PCB_HTM,
from there after, the context switch will save and restore the 3 SPRs that are
associated with HTM.

This also adds an HTM example and testcase for Powerpc64. HTM is a hardware feature
that showed up initially on POWER8, so, you need at least this hardware version
to exploit this feature.

The test case needs gcc6 or older to be compiled, otherwise, there are
instructions that are not recognized and the build fails.